fix: email confirmation email resending
