+++ web/src/app/confirm-email/pending/_test/pending.test.tsx
@@ -1,40 +1,41 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 
-vi.stubGlobal("fetch", vi.fn().mockResolvedValue({ ok: false, json: async () => ({ error: "Failed" }) }));
+vi.mock("server-only", () => ({}));
+vi.mock("@/lib/auth/userFromToken", () => ({
+  getUser: vi.fn().mockResolvedValue({
+    email: "user@example.com",
+  }),
+}))
 
-Object.defineProperty(window, "location", {
-  value: { href: "" },
-  writable: true,
-});
+vi.stubGlobal("fetch", vi.fn().mockResolvedValue({ ok: false, json: async () => ({ error: "Failed" }) }))
 
 import ConfirmEmailPendingPage from "../page";
 
 beforeEach(() => {
   vi.clearAllMocks();
   window.location.href = "";
-  localStorage.setItem("pending_confirmation_email", "test@test.com");
   vi.mocked(fetch).mockResolvedValue({ ok: false, json: async () => ({ error: "Failed" }) } as any);
 });
 
 describe("ConfirmEmailPendingPage", () => {
-  it("renders heading", () => {
-    render(<ConfirmEmailPendingPage />);
+  it("renders heading", async () => {
+    render(await ConfirmEmailPendingPage());
     expect(screen.getByText(/check your inbox/i)).toBeInTheDocument();
   });
 
-  it("renders resend button", () => {
-    render(<ConfirmEmailPendingPage />);
+  it("renders resend button", async () => {
+    render(await ConfirmEmailPendingPage());
     expect(screen.getByRole("button", { name: /resend confirmation email/i })).toBeInTheDocument();
   });
 
-  it("calls resend API with email from localStorage", async () => {
-    render(<ConfirmEmailPendingPage />);
+  it("calls resend API with email", async () => {
+    render(await ConfirmEmailPendingPage());
     fireEvent.click(screen.getByRole("button", { name: /resend confirmation email/i }));
     await waitFor(() => {
       expect(fetch).toHaveBeenCalledWith("/api/auth/resend", expect.objectContaining({
         method: "POST",
-        body: JSON.stringify({ email: "test@test.com" }),
+        body: JSON.stringify({ email: "user@example.com" }),
       }));
     });
   });
@@ -45,7 +46,7 @@ describe("ConfirmEmailPendingPage", () => {
       json: async () => ({ confirmURL: "https://example.com/verify" }),
     } as any);
 
-    render(<ConfirmEmailPendingPage />);
+    render(await ConfirmEmailPendingPage());
     fireEvent.click(screen.getByRole("button", { name: /resend confirmation email/i }));
     await waitFor(() => {
       expect(screen.getByText(/✓ email sent/i)).toBeInTheDocument();
@@ -54,7 +55,7 @@ describe("ConfirmEmailPendingPage", () => {
 
   it("disables button while loading", async () => {
     vi.mocked(fetch).mockReturnValue(new Promise(() => {}) as any);
-    render(<ConfirmEmailPendingPage />);
+    render(await ConfirmEmailPendingPage());
     fireEvent.click(screen.getByRole("button", { name: /resend confirmation email/i }));
     await waitFor(() => {
       expect(screen.getByRole("button", { name: /sending/i })).toBeDisabled();
